Better error handling in the SQL level
Throwing errors using Haskell exceptions makes more sense if our monad
stack has IO in it. Moreover, it's the correct way to interact with things
like PGS.withTransaction anyway, since they rely on exceptions to decide if
a transaction shall be rolled back or committed.

For now we only check and handle those invariants whose violation implies
serious issues in the program logic (like failure to update a post version).
Incorrect removed count (as in some other functions) is not as severe, so
we'll handle it later when we have proper logging around.